Add ConfigureCounter component and update Counter functionality
A new ConfigureCounter component is added to handle counter configuration. The Counter component is modified to implement a memoized version of whether the initial count is prime, in order to optimize the performance. Furthermore, it also includes a CounterHistory component to keep track of counter changes. Refactoring is also done to reduce duplication in the App component.